+++ recipes/rhel.rb
@@ -2,7 +2,7 @@
 # Cookbook Name:: annoyances
 # Recipe:: rhel
 #
-# Copyright 2012, Opscode, Inc.
+# Copyright 2012-2013, Opscode, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,15 +21,26 @@
 execute("iptables -F") { ignore_failure true }.run_action(:run)
 
 #turn off SELinux
-execute("setenforce 0") { ignore_failure true }.run_action(:run)
+if Mixlib::ShellOut.new("getenforce").run_command.stdout != "Disabled\n" then
+  execute("setenforce 0") { ignore_failure true }.run_action(:run)
+end
 
-execute"rpm --nodeps -e httpd" do
-  ignore_failure true
-  not_if do
-    node['recipes'].include?("apache2") ||
-      run_context.loaded_recipe?("apache2")
+#uninstall httpd
+if Mixlib::ShellOut.new("rpm -q httpd").run_command.status.success? then
+  execute "rpm --nodeps -e httpd" do
+    ignore_failure true
+    not_if do
+      node.recipe?("apache2")
+    end
   end
 end
 
 #remove any .bash_logout
 file("/root/.bash_logout") { action :delete }
+
+#disable the pile of desktop services that get turned on by default
+%w{autofs avahi-daemon bluetooth cpuspeed cups gpm haldaemon messagebus}.each do |svc|
+  service svc do
+    action [:stop, :disable]
+  end
+end